pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faithcheck"
version = "0.1.0"
description = "Causal CoT step faithfulness evaluation harness"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.4",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.4",
]

[project.scripts]
faithcheck = "faithcheck.cli.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true
python_version = "3.10"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]
